+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/accountinfo/AccountInfo.java
@@ -3,8 +3,9 @@ package com.concordium.sdk.responses.accountinfo;
 import com.concordium.sdk.responses.accountinfo.credential.Credential;
 import com.concordium.sdk.serializing.JsonMapper;
 import com.concordium.sdk.transactions.AccountAddress;
-import com.concordium.sdk.transactions.GTUAmount;
+import com.concordium.sdk.transactions.CCDAmount;
 import com.concordium.sdk.transactions.Index;
+import com.concordium.sdk.types.Nonce;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -24,13 +25,13 @@ public final class AccountInfo {
      */
     private final AccountAddress accountAddress;
     /**
-     * The next available nonce for the account.
+     * The nonce for the account.
      */
-    private final int accountNonce;
+    private final Nonce accountNonce;
     /**
      * The amount of CCD available for the account.
      */
-    private final GTUAmount accountAmount;
+    private final CCDAmount accountAmount;
     /**
      * The account threshold for this account i.e., how
      * many credentials that needs to sign transactions for this account.
@@ -71,8 +72,8 @@ public final class AccountInfo {
 
     @JsonCreator
     AccountInfo(@JsonProperty("accountAddress") String accountAddress,
-                @JsonProperty("accountNonce") int accountNonce,
-                @JsonProperty("accountAmount") GTUAmount accountAmount,
+                @JsonProperty("accountNonce") Nonce accountNonce,
+                @JsonProperty("accountAmount") CCDAmount accountAmount,
                 @JsonProperty("accountThreshold") int accountThreshold,
                 @JsonProperty("accountEncryptionKey") String accountEncryptionKey,
                 @JsonProperty("accountIndex") int accountIndex,