+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/accountinfo/ScheduledRelease.java
@@ -1,6 +1,6 @@
 package com.concordium.sdk.responses.accountinfo;
 
-import com.concordium.sdk.transactions.GTUAmount;
+import com.concordium.sdk.transactions.CCDAmount;
 import com.concordium.sdk.types.Timestamp;
 
 import com.concordium.sdk.transactions.Hash;
@@ -15,12 +15,12 @@ import java.util.List;
 @ToString
 public final class ScheduledRelease {
     private final Timestamp timestamp;
-    private final GTUAmount amount;
+    private final CCDAmount amount;
     private final List<Hash> transactions;
 
     @JsonCreator
     ScheduledRelease(@JsonProperty("timestamp") Timestamp timestamp,
-                     @JsonProperty("amount") GTUAmount amount,
+                     @JsonProperty("amount") CCDAmount amount,
                      @JsonProperty("transactions") List<Hash> transactions) {
         this.timestamp = timestamp;
         this.amount = amount;