+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/AddAnonymityRevokerUpdates.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.responses.blocksummary;
 
+import com.concordium.sdk.types.Nonce;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -9,23 +10,17 @@ import java.util.List;
 
 @Getter
 @ToString
-public class TransactionFeeDistribution {
-    private final double gasAccount;
-    private final double baker;
+public final class AddAnonymityRevokerUpdates {
     private final int threshold;
     private final List<Integer> authorizedKeys;
-    private final int nextSequenceNumber;
+    private final Nonce nextSequenceNumber;
     private final List<Object> queue;
 
     @JsonCreator
-    TransactionFeeDistribution(@JsonProperty("gasAccount") double gasAccount,
-                               @JsonProperty("baker") double baker,
-                               @JsonProperty("threshold") int threshold,
+    AddAnonymityRevokerUpdates(@JsonProperty("threshold") int threshold,
                                @JsonProperty("authorizedKeys") List<Integer> authorizedKeys,
-                               @JsonProperty("nextSequenceNumber") int nextSequenceNumber,
+                               @JsonProperty("nextSequenceNumber") Nonce nextSequenceNumber,
                                @JsonProperty("queue") List<Object> queue) {
-        this.gasAccount = gasAccount;
-        this.baker = baker;
         this.threshold = threshold;
         this.authorizedKeys = authorizedKeys;
         this.nextSequenceNumber = nextSequenceNumber;