+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/AddIdentityProviderUpdates.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.responses.blocksummary;
 
+import com.concordium.sdk.types.Nonce;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -9,16 +10,16 @@ import java.util.List;
 
 @Getter
 @ToString
-public final class BakerStakeThreshold {
+public final class AddIdentityProviderUpdates {
     private final int threshold;
     private final List<Integer> authorizedKeys;
-    private final int nextSequenceNumber;
+    private final Nonce nextSequenceNumber;
     private final List<Object> queue;
 
     @JsonCreator
-    BakerStakeThreshold(@JsonProperty("threshold") int threshold,
+    AddIdentityProviderUpdates(@JsonProperty("threshold") int threshold,
                                @JsonProperty("authorizedKeys") List<Integer> authorizedKeys,
-                               @JsonProperty("nextSequenceNumber") int nextSequenceNumber,
+                               @JsonProperty("nextSequenceNumber") Nonce nextSequenceNumber,
                                @JsonProperty("queue") List<Object> queue) {
         this.threshold = threshold;
         this.authorizedKeys = authorizedKeys;