+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/BakerStakeThresholdUpdates.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.responses.blocksummary;
 
+import com.concordium.sdk.types.Nonce;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -9,17 +10,17 @@ import java.util.List;
 
 @Getter
 @ToString
-public final class FoundationAccount {
+public final class BakerStakeThresholdUpdates {
     private final int threshold;
     private final List<Integer> authorizedKeys;
-    private final int nextSequenceNumber;
+    private final Nonce nextSequenceNumber;
     private final List<Object> queue;
 
     @JsonCreator
-    public FoundationAccount(@JsonProperty("threshold") int threshold,
-                             @JsonProperty("authorizedKeys") List<Integer> authorizedKeys,
-                             @JsonProperty("nextSequenceNumber") int nextSequenceNumber,
-                             @JsonProperty("queue") List<Object> queue) {
+    BakerStakeThresholdUpdates(@JsonProperty("threshold") int threshold,
+                               @JsonProperty("authorizedKeys") List<Integer> authorizedKeys,
+                               @JsonProperty("nextSequenceNumber") Nonce nextSequenceNumber,
+                               @JsonProperty("queue") List<Object> queue) {
         this.threshold = threshold;
         this.authorizedKeys = authorizedKeys;
         this.nextSequenceNumber = nextSequenceNumber;