+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/ChainParameters.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.responses.blocksummary;
 
+import com.concordium.sdk.transactions.CCDAmount;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -8,31 +9,31 @@ import lombok.ToString;
 @Getter
 @ToString
 public final class ChainParameters {
-    private final String minimumThresholdForBaking;
+    private final CCDAmount minimumThresholdForBaking;
     private final RewardParameters rewardParameters;
-    private final MicroGTUPerEuro microGTUPerEuro;
+    private final MicroGTUPerEuroUpdates microGTUPerEuroUpdates;
     private final int foundationAccountIndex;
     private final int accountCreationLimit;
     private final int bakerCooldownEpochs;
     private final double electionDifficulty;
-    private final EuroPerEnergy euroPerEnergy;
+    private final EuroPerEnergyUpdates euroPerEnergyUpdates;
 
     @JsonCreator
-    ChainParameters(@JsonProperty("minimumThresholdForBaking") String minimumThresholdForBaking,
+    ChainParameters(@JsonProperty("minimumThresholdForBaking") CCDAmount minimumThresholdForBaking,
                            @JsonProperty("rewardParameters") RewardParameters rewardParameters,
-                           @JsonProperty("microGTUPerEuro") MicroGTUPerEuro microGTUPerEuro,
+                           @JsonProperty("microGTUPerEuro") MicroGTUPerEuroUpdates microGTUPerEuroUpdates,
                            @JsonProperty("foundationAccountIndex") int foundationAccountIndex,
                            @JsonProperty("accountCreationLimit") int accountCreationLimit,
                            @JsonProperty("bakerCooldownEpochs") int bakerCooldownEpochs,
                            @JsonProperty("electionDifficulty") double electionDifficulty,
-                           @JsonProperty("euroPerEnergy") EuroPerEnergy euroPerEnergy) {
+                           @JsonProperty("euroPerEnergy") EuroPerEnergyUpdates euroPerEnergyUpdates) {
         this.minimumThresholdForBaking = minimumThresholdForBaking;
         this.rewardParameters = rewardParameters;
-        this.microGTUPerEuro = microGTUPerEuro;
+        this.microGTUPerEuroUpdates = microGTUPerEuroUpdates;
         this.foundationAccountIndex = foundationAccountIndex;
         this.accountCreationLimit = accountCreationLimit;
         this.bakerCooldownEpochs = bakerCooldownEpochs;
         this.electionDifficulty = electionDifficulty;
-        this.euroPerEnergy = euroPerEnergy;
+        this.euroPerEnergyUpdates = euroPerEnergyUpdates;
     }
 }