+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/Finalizer.java
@@ -5,17 +5,19 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.ToString;
 
+import java.math.BigInteger;
+
 @Getter
 @ToString
 public final class Finalizer {
     private final int bakerId;
-    private final Object weight;
+    private final BigInteger weight;
     private final boolean signed;
 
 
     @JsonCreator
     Finalizer(@JsonProperty("bakerId") int bakerId,
-                     @JsonProperty("weight") Object weight,
+                     @JsonProperty("weight") BigInteger weight,
                      @JsonProperty("signed") boolean signed) {
         this.bakerId = bakerId;
         this.weight = weight;