+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/GasRewardsUpdates.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.responses.blocksummary;
 
+import com.concordium.sdk.types.Nonce;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -9,13 +10,13 @@ import java.util.List;
 
 @Getter
 @ToString
-public final class GasRewards2 {
-    private final int nextSequenceNumber;
+public final class GasRewardsUpdates {
+    private final Nonce nextSequenceNumber;
     private final List<Object> queue;
 
     @JsonCreator
-    GasRewards2(@JsonProperty("nextSequenceNumber") int nextSequenceNumber,
-                @JsonProperty("queue") List<Object> queue) {
+    GasRewardsUpdates(@JsonProperty("nextSequenceNumber") Nonce nextSequenceNumber,
+                      @JsonProperty("queue") List<Object> queue) {
         this.nextSequenceNumber = nextSequenceNumber;
         this.queue = queue;
     }