+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/MintDistributionUpdates.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.responses.blocksummary;
 
+import com.concordium.sdk.types.Nonce;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -9,23 +10,23 @@ import java.util.List;
 
 @Getter
 @ToString
-public final class MintDistribution {
+public final class MintDistributionUpdates {
     private final double mintPerSlot;
     private final double bakingReward;
     private final double finalizationReward;
     private final int threshold;
     private final List<Integer> authorizedKeys;
-    private final int nextSequenceNumber;
+    private final Nonce nextSequenceNumber;
     private final List<Object> queue;
 
     @JsonCreator
-    MintDistribution(@JsonProperty("mintPerSlot") double mintPerSlot,
-                     @JsonProperty("bakingReward") double bakingReward,
-                     @JsonProperty("finalizationReward") double finalizationReward,
-                     @JsonProperty("threshold") int threshold,
-                     @JsonProperty("authorizedKeys") List<Integer> authorizedKeys,
-                     @JsonProperty("nextSequenceNumber") int nextSequenceNumber,
-                     @JsonProperty("queue") List<Object> queue) {
+    MintDistributionUpdates(@JsonProperty("mintPerSlot") double mintPerSlot,
+                            @JsonProperty("bakingReward") double bakingReward,
+                            @JsonProperty("finalizationReward") double finalizationReward,
+                            @JsonProperty("threshold") int threshold,
+                            @JsonProperty("authorizedKeys") List<Integer> authorizedKeys,
+                            @JsonProperty("nextSequenceNumber") Nonce nextSequenceNumber,
+                            @JsonProperty("queue") List<Object> queue) {
         this.mintPerSlot = mintPerSlot;
         this.bakingReward = bakingReward;
         this.finalizationReward = finalizationReward;