+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/transactionstatus/RejectReasonAmountTooLarge.java
@@ -1,6 +1,6 @@
 package com.concordium.sdk.responses.transactionstatus;
 
-import com.concordium.sdk.transactions.GTUAmount;
+import com.concordium.sdk.transactions.CCDAmount;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -14,7 +14,7 @@ import java.util.Map;
 public class RejectReasonAmountTooLarge extends RejectReason {
 
     private final AbstractAccount account;
-    private final GTUAmount amount;
+    private final CCDAmount amount;
 
     @JsonCreator
     RejectReasonAmountTooLarge(@JsonProperty("contents") List<Object> contents) {
@@ -23,7 +23,7 @@ public class RejectReasonAmountTooLarge extends RejectReason {
         }
         try {
             this.account = AbstractAccount.parseAccount((Map<String, Object>) contents.get(0));
-            this.amount = GTUAmount.fromMicro((String) contents.get(1));
+            this.amount = CCDAmount.fromMicro((String) contents.get(1));
         } catch (ClassCastException e) {
             throw new IllegalArgumentException("Cannot parse AmountTooLarge. Unexpected type.", e);
         }