+++ concordium-sdk/src/test/java/com/concordium/sdk/transactions/TransferTest.java
@@ -1,6 +1,7 @@
 package com.concordium.sdk.transactions;
 
 import com.concordium.sdk.crypto.ed25519.ED25519SecretKey;
+import com.concordium.sdk.types.Nonce;
 import com.concordium.sdk.types.UInt64;
 import lombok.SneakyThrows;
 import lombok.val;
@@ -18,11 +19,11 @@ public class TransferTest {
     public void testCreateTransfer() {
         val transfer = Transfer.createNew(
                         AccountAddress.from("3hYXYEPuGyhFcVRhSk2cVgKBhzVcAryjPskYk4SecpwGnoHhuM"),
-                        GTUAmount.fromMicro(17))
+                        CCDAmount.fromMicro(17))
                 .withHeader(TransactionHeader
                         .builder()
                         .sender(AccountAddress.from("3JwD2Wm3nMbsowCwb1iGEpnt47UQgdrtnq2qT6opJc3z2AgCrc"))
-                        .accountNonce(UInt64.from(78910))
+                        .accountNonce(Nonce.from(78910))
                         .expiry(UInt64.from(123456))
                         .build())
                 .signWith(